seahawks_client/requirements.txt
requests==2.31.0
psutil==5.9.5
python-nmap==0.7.1
flask==2.3.3
